Added very basic functionality with Firebase. NewWorkOrderViewController will allow you to save the work order description and priority to the Firebase database.
Next steps, add some user authentication and auto-populate the "Entered by" field and "Date Entered" fields of the new work order view controller. Need to also get a separate dummy database of Assets in Firebase and allow the user to select from that when writing a new work order.
Also need to pull data from Firebase to populate the TableViewController cells in the "View Work Orders" part of the application